Tidy up example app project

+++ projects/example-app/src/app/app.component.ts
@@ -1,22 +1,24 @@
 import {AfterViewInit, Component, ElementRef, ViewChild} from '@angular/core';
-import {PetriflowCanvasService} from '../../petriflow-svg/src/lib/petriflow-canvas.service';
+import {
+    CanvasMode,
+    PetriflowCanvasConfigurationService,
+    PetriflowCanvasFactoryService,
+    PetriflowCanvasService
+} from 'petriflow-svg';
 import {MatToolbar} from '@angular/material/toolbar';
-import {PetriflowCanvasFactoryService} from '../../petriflow-svg/src/lib/factories/petriflow-canvas-factory.service';
-import {CanvasMode} from '../../petriflow-svg/src/lib/canvas-mode';
-import {PetriflowCanvasConfigurationService} from '../../petriflow-svg/src/lib/services/petriflow-canvas-configuration.service';
 import {MatDialog} from '@angular/material/dialog';
 import {PetriflowInfoDialogComponent} from './petriflow-info-dialog/petriflow-info-dialog.component';
 
 @Component({
-    selector: 'nab-root',
+    selector: 'pf-app-root',
     templateUrl: './app.component.html',
     styleUrls: ['./app.component.scss']
 })
 export class AppComponent implements AfterViewInit {
 
-    @ViewChild(MatToolbar) toolbar: MatToolbar;
-    @ViewChild('canvasComponent') canvasComponent: ElementRef;
-    public _mode: CanvasMode;
+    @ViewChild(MatToolbar) toolbar: MatToolbar | undefined;
+    @ViewChild('canvasComponent') canvasComponent: ElementRef | undefined;
+    public _mode: CanvasMode | undefined;
 
     constructor(private _petriflowCanvasService: PetriflowCanvasService, private _petriflowFactoryService: PetriflowCanvasFactoryService,
                 private _petriflowConfigService: PetriflowCanvasConfigurationService, public dialog: MatDialog) {
@@ -24,21 +26,25 @@ export class AppComponent implements AfterViewInit {
     }
 
     ngAfterViewInit(): void {
+        if (!this._petriflowCanvasService.canvas)
+            throw new Error("Petriflow SVG canvas does not exists!");
         this._petriflowCanvasService.canvas.svg.onclick = (e) => {
             this.addTransition(e);
             this.addPlace(e);
         };
+        if (!this.toolbar)
+            throw new Error("MatToolbar could not be found!");
         this._petriflowConfigService.addCanvasEvent(this._petriflowCanvasService.canvas.svg, this.toolbar);
         this.toolbar._elementRef.nativeElement.onmouseenter = () => {
             this._petriflowConfigService.deleteClipboard();
         };
     }
 
-    private addTransition($event): void {
+    private addTransition($event: MouseEvent): void {
         if (this._petriflowConfigService.mode === CanvasMode.CREATE_TRANSITION) {
             const offset = this._petriflowCanvasService.getPanZoomOffset();
-            const transition = this._petriflowFactoryService.createTransition(new DOMPoint(($event.offsetX - offset.x) / offset.scale,
-                ($event.offsetY - offset.y) / offset.scale));
+            const transition = this._petriflowFactoryService.createTransition(new DOMPoint(($event.offsetX - (offset?.x ?? 0)) / (offset?.scale ?? 1),
+                ($event.offsetY - (offset?.y ?? 0)) / (offset?.scale ?? 1)));
             this._petriflowConfigService.addTransitionEvents(transition);
 
         }
@@ -47,26 +53,26 @@ export class AppComponent implements AfterViewInit {
     private addPlace(e: MouseEvent) {
         if (this._petriflowConfigService.mode === CanvasMode.CREATE_PLACE) {
             const offset = this._petriflowCanvasService.getPanZoomOffset();
-            const place = this._petriflowFactoryService.createPlace(0, new DOMPoint((e.offsetX - offset.x) / offset.scale, (e.offsetY - offset.y) / offset.scale));
+            const place = this._petriflowFactoryService.createPlace(0, new DOMPoint((e.offsetX - (offset?.x ?? 0)) / (offset?.scale ?? 1), (e.offsetY - (offset?.y ?? 0)) / (offset?.scale ?? 1)));
             this._petriflowConfigService.addPlaceEvents(place);
         }
     }
 
     disablePreviousArcMode() {
         if (this._petriflowFactoryService.arcLine) {
-            this._petriflowCanvasService.canvas.container.removeChild(this._petriflowFactoryService.arcLine);
+            this._petriflowCanvasService.canvas?.container.removeChild(this._petriflowFactoryService.arcLine);
             this._petriflowFactoryService.source = undefined;
             this._petriflowFactoryService.arcLine = undefined;
         }
     }
 
-    changeCanvasMode(mode: CanvasMode, panzoomEnabled = true, cursor?: string) {
+    changeCanvasMode(mode: CanvasMode, panzoomEnabled = true) {
         this.disablePreviousArcMode();
         this._petriflowConfigService.mode = mode;
-        if (panzoomEnabled && this._petriflowCanvasService.panzoom.isPaused()) {
+        if (panzoomEnabled && this._petriflowCanvasService.panzoom?.isPaused()) {
             this._petriflowCanvasService.panzoom.resume();
-        } else if (!panzoomEnabled && !this._petriflowCanvasService.panzoom.isPaused()) {
-            this._petriflowCanvasService.panzoom.pause();
+        } else if (!panzoomEnabled && !this._petriflowCanvasService.panzoom?.isPaused()) {
+            this._petriflowCanvasService.panzoom?.pause();
         }
     }
 
@@ -75,8 +81,8 @@ export class AppComponent implements AfterViewInit {
     }
 
     resetPanZoom() {
-        this._petriflowCanvasService.panzoom.moveTo(0, 0);
-        this._petriflowCanvasService.panzoom.zoomAbs(0, 0, 1);
+        this._petriflowCanvasService.panzoom?.moveTo(0, 0);
+        this._petriflowCanvasService.panzoom?.zoomAbs(0, 0, 1);
     }
 
     public get canvasMode(): typeof CanvasMode {
